requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
redis==5.1.0
bcrypt==4.0.1